python/requirements.txt
python-docx>=1.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pytest>=8.0.0
httpx>=0.27.0
